# packages/backend/src/generated/swagger.json
{
	"components": {
...
		"schemas": {
			"ApiErrorPayload": {
				"properties": {
					"error": {
						"properties": {
							"data": {
								"description": "Optional data containing details of the error"
							},
							"message": {
								"type": "string",
								"description": "A friendly message summarising the error"
							},
							"name": {
								"type": "string",
								"description": "Unique name for the type of error"
							},
							"statusCode": {
								"type": "number",
								"format": "integer",
								"description": "HTTP status code"
							}
						},
						"required": [
							"name",
							"statusCode"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"error"
						],
						"nullable": false
					}
				},
				"required": [
					"error",
					"status"
				],
				"type": "object",
				"description": "The Error object is returned from the api any time there is an error.\nThe message contains"
			},
...
			"ApiSuccessEmpty": {
				"properties": {
					"results": {},
					"status": {
						"type": "string",
						"enum": [
							"ok"
						],
						"nullable": false
					}
				},
				"required": [
					"status"
				],
				"type": "object"
			},
			"Pick_Explore.SummaryExploreFields_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"groupLabel": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"label",
					"tags"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"SummaryExtraFields": {
				"properties": {
					"databaseName": {
						"type": "string"
					},
					"schemaName": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"databaseName",
					"schemaName"
				],
				"type": "object"
			},
			"InlineErrorType": {
				"enum": [
					"METADATA_PARSE_ERROR",
					"NO_DIMENSIONS_FOUND"
				],
				"type": "string"
			},
			"InlineError": {
				"properties": {
					"message": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/InlineErrorType"
					}
				},
				"required": [
					"message",
					"type"
				],
				"type": "object"
			},
			"Pick_ExploreError.SummaryExploreErrorFields_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"groupLabel": {
						"type": "string"
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/InlineError"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"label",
					"tags",
					"errors"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Partial_SummaryExtraFields_": {
				"properties": {
					"description": {
						"type": "string"
					},
					"schemaName": {
						"type": "string"
					},
					"databaseName": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"SummaryExplore": {
				"anyOf": [
					{
						"allOf": [
							{
								"$ref": "#/components/schemas/Pick_Explore.SummaryExploreFields_"
							},
							{
								"$ref": "#/components/schemas/SummaryExtraFields"
							}
						]
					},
					{
						"allOf": [
							{
								"$ref": "#/components/schemas/Pick_ExploreError.SummaryExploreErrorFields_"
							},
							{
								"$ref": "#/components/schemas/Partial_SummaryExtraFields_"
							}
						]
					}
				]
			},
			"ApiExploresResults": {
				"items": {
					"$ref": "#/components/schemas/SummaryExplore"
				},
				"type": "array"
			},
			"DbtModelJoinType": {
				"type": "string",
				"enum": [
					"inner",
					"full",
					"left",
					"right"
				]
			},
			"Pick_ExploreJoin.table-or-sqlOn-or-type-or-hidden_": {
				"properties": {
					"table": {
						"type": "string"
					},
					"sqlOn": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/DbtModelJoinType"
					},
					"hidden": {
						"type": "boolean"
					}
				},
				"required": [
					"table",
					"sqlOn"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"CompiledExploreJoin": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Pick_ExploreJoin.table-or-sqlOn-or-type-or-hidden_"
					},
					{
						"properties": {
							"compiledSqlOn": {
								"type": "string"
							}
						},
						"required": [
							"compiledSqlOn"
						],
						"type": "object"
					}
				]
			},
			"OrderFieldsByStrategy": {
				"enum": [
					"LABEL",
					"INDEX"
				],
				"type": "string"
			},
			"Record_string.string-or-string-Array_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"TableBase": {
				"properties": {
					"requiredAttributes": {
						"$ref": "#/components/schemas/Record_string.string-or-string-Array_"
					},
					"hidden": {
						"type": "boolean"
					},
					"sqlWhere": {
						"type": "string"
					},
					"groupLabel": {
						"type": "string"
					},
					"orderFieldsBy": {
						"$ref": "#/components/schemas/OrderFieldsByStrategy"
					},
					"sqlTable": {
						"type": "string"
					},
					"schema": {
						"type": "string"
					},
					"database": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"sqlTable",
					"schema",
					"database",
					"label",
					"name"
				],
				"type": "object"
			},
			"Record_string.CompiledDimension_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"Record_string.CompiledMetric_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"Record_string.LineageNodeDependency-Array_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"LineageGraph": {
				"$ref": "#/components/schemas/Record_string.LineageNodeDependency-Array_"
			},
			"SourcePosition": {
				"properties": {
					"character": {
						"type": "number",
						"format": "double"
					},
					"line": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"character",
					"line"
				],
				"type": "object"
			},
			"Source": {
				"properties": {
					"content": {
						"type": "string"
					},
					"highlight": {
						"properties": {
							"end": {
								"$ref": "#/components/schemas/SourcePosition"
							},
							"start": {
								"$ref": "#/components/schemas/SourcePosition"
							}
						},
						"required": [
							"end",
							"start"
						],
						"type": "object"
					},
					"range": {
						"properties": {
							"end": {
								"$ref": "#/components/schemas/SourcePosition"
							},
							"start": {
								"$ref": "#/components/schemas/SourcePosition"
							}
						},
						"required": [
							"end",
							"start"
						],
						"type": "object"
					},
					"path": {
						"type": "string"
					}
				},
				"required": [
					"content",
					"range",
					"path"
				],
				"type": "object"
			},
			"CompiledTable": {
				"allOf": [
					{
						"$ref": "#/components/schemas/TableBase"
					},
					{
						"properties": {
							"source": {
								"$ref": "#/components/schemas/Source"
							},
							"lineageGraph": {
								"$ref": "#/components/schemas/LineageGraph"
							},
							"metrics": {
								"$ref": "#/components/schemas/Record_string.CompiledMetric_"
							},
							"dimensions": {
								"$ref": "#/components/schemas/Record_string.CompiledDimension_"
							}
						},
						"required": [
							"lineageGraph",
							"metrics",
							"dimensions"
						],
						"type": "object"
					}
				]
			},
			"SupportedDbtAdapter": {
				"enum": [
					"bigquery",
					"databricks",
					"snowflake",
					"redshift",
					"postgres",
					"trino"
				],
				"type": "string"
			},
			"Explore": {
				"properties": {
					"sqlPath": {
						"type": "string"
					},
					"ymlPath": {
						"type": "string"
					},
					"warehouse": {
						"type": "string"
					},
					"targetDatabase": {
						"$ref": "#/components/schemas/SupportedDbtAdapter"
					},
					"tables": {
						"properties": {},
						"additionalProperties": {
							"$ref": "#/components/schemas/CompiledTable"
						},
						"type": "object"
					},
					"joinedTables": {
						"items": {
							"$ref": "#/components/schemas/CompiledExploreJoin"
						},
						"type": "array"
					},
					"baseTable": {
						"type": "string"
					},
					"groupLabel": {
						"type": "string"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"label": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"targetDatabase",
					"tables",
					"joinedTables",
					"baseTable",
					"tags",
					"label",
					"name"
				],
				"type": "object"
			},
			"ApiExploreResults": {
				"$ref": "#/components/schemas/Explore"
			},
			"ApiCompiledQueryResults": {
				"type": "string"
			},
			"FieldId": {
				"type": "string"
			},
			"FilterGroup": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/OrFilterGroup"
					},
					{
						"$ref": "#/components/schemas/AndFilterGroup"
					}
				]
			},
			"FieldTarget": {
				"properties": {
					"fieldId": {
						"type": "string"
					}
				},
				"required": [
					"fieldId"
				],
				"type": "object"
			},
			"ConditionalOperator": {
				"enum": [
					"isNull",
					"notNull",
					"equals",
					"notEquals",
					"startsWith",
					"endsWith",
					"include",
					"doesNotInclude",
					"lessThan",
					"lessThanOrEqual",
					"greaterThan",
					"greaterThanOrEqual",
					"inThePast",
					"notInThePast",
					"inTheNext",
					"inTheCurrent",
					"inBetween"
				],
				"type": "string"
			},
			"FilterRule": {
				"properties": {
					"values": {
						"items": {},
						"type": "array"
					},
					"operator": {
						"$ref": "#/components/schemas/ConditionalOperator"
					},
					"id": {
						"type": "string"
					},
					"target": {
						"$ref": "#/components/schemas/FieldTarget"
					},
					"settings": {},
					"disabled": {
						"type": "boolean"
					}
				},
				"required": [
					"operator",
					"id",
					"target"
				],
				"type": "object",
				"additionalProperties": true
			},
			"FilterGroupItem": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/FilterGroup"
					},
					{
						"$ref": "#/components/schemas/FilterRule"
					}
				]
			},
			"OrFilterGroup": {
				"properties": {
					"or": {
						"items": {
							"$ref": "#/components/schemas/FilterGroupItem"
						},
						"type": "array"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"or",
					"id"
				],
				"type": "object"
			},
			"AndFilterGroup": {
				"properties": {
					"and": {
						"items": {
							"$ref": "#/components/schemas/FilterGroupItem"
						},
						"type": "array"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"and",
					"id"
				],
				"type": "object"
			},
			"Filters": {
				"properties": {
					"tableCalculations": {
						"$ref": "#/components/schemas/FilterGroup"
					},
					"metrics": {
						"$ref": "#/components/schemas/FilterGroup"
					},
					"dimensions": {
						"$ref": "#/components/schemas/FilterGroup"
					}
				},
				"type": "object"
			},
			"SortField": {
				"properties": {
					"descending": {
						"type": "boolean"
					},
					"fieldId": {
						"type": "string"
					}
				},
				"required": [
					"descending",
					"fieldId"
				],
				"type": "object"
			},
			"CustomFormatType": {
				"enum": [
					"default",
					"percent",
					"currency",
					"number",
					"id"
				],
				"type": "string"
			},
			"NumberSeparator": {
				"enum": [
					"default",
					"commaPeriod",
					"spacePeriod",
					"periodComma",
					"noSeparatorPeriod"
				],
				"type": "string"
			},
			"Compact": {
				"enum": [
					"thousands",
					"millions",
					"billions",
					"trillions"
				],
				"type": "string"
			},
			"CompactOrAlias": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/Compact"
					},
					{
						"type": "string",
						"enum": [
							"K",
							"thousand",
							"M",
							"million",
							"B",
							"billion",
							"T",
							"trillion"
						]
					}
				]
			},
			"CustomFormat": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/CustomFormatType"
					},
					"round": {
						"type": "number",
						"format": "double"
					},
					"separator": {
						"$ref": "#/components/schemas/NumberSeparator"
					},
					"currency": {
						"type": "string"
					},
					"compact": {
						"$ref": "#/components/schemas/CompactOrAlias"
					},
					"prefix": {
						"type": "string"
					},
					"suffix": {
						"type": "string"
					}
				},
				"required": [
					"type"
				],
				"type": "object",
				"additionalProperties": true
			},
			"TableCalculation": {
				"properties": {
					"format": {
						"$ref": "#/components/schemas/CustomFormat"
					},
					"sql": {
						"type": "string"
					},
					"displayName": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"index": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"sql",
					"displayName",
					"name"
				],
				"type": "object"
			},
			"MetricType": {
				"enum": [
					"percentile",
					"average",
					"count",
					"count_distinct",
					"sum",
					"min",
					"max",
					"number",
					"median",
					"string",
					"date",
					"timestamp",
					"boolean"
				],
				"type": "string"
			},
			"Format": {
				"enum": [
					"km",
					"mi",
					"usd",
					"gbp",
					"eur",
					"id",
					"percent"
				],
				"type": "string"
			},
			"MetricFilterRule": {
				"properties": {
					"values": {
						"items": {},
						"type": "array"
					},
					"operator": {
						"$ref": "#/components/schemas/ConditionalOperator"
					},
					"id": {
						"type": "string"
					},
					"target": {
						"properties": {
							"fieldRef": {
								"type": "string"
							}
						},
						"required": [
							"fieldRef"
						],
						"type": "object"
					},
					"settings": {},
					"disabled": {
						"type": "boolean"
					}
				},
				"required": [
					"operator",
					"id",
					"target"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AdditionalMetric": {
				"properties": {
					"label": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/MetricType"
					},
					"description": {
						"type": "string"
					},
					"sql": {
						"type": "string"
					},
					"hidden": {
						"type": "boolean"
					},
					"round": {
						"type": "number",
						"format": "double"
					},
					"compact": {
						"$ref": "#/components/schemas/CompactOrAlias"
					},
					"format": {
						"$ref": "#/components/schemas/Format"
					},
					"table": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"index": {
						"type": "number",
						"format": "double"
					},
					"filters": {
						"items": {
							"$ref": "#/components/schemas/MetricFilterRule"
						},
						"type": "array"
					},
					"baseDimensionName": {
						"type": "string"
					},
					"uuid": {
						"type": "string",
						"nullable": true
					},
					"percentile": {
						"type": "number",
						"format": "double"
					},
					"formatOptions": {
						"$ref": "#/components/schemas/CustomFormat"
					}
				},
				"required": [
					"type",
					"sql",
					"table",
					"name"
				],
				"type": "object",
				"additionalProperties": true
			},
			"BinType": {
				"enum": [
					"fixed_number",
					"fixed_width",
					"custom_range"
				],
				"type": "string"
			},
			"BinRange": {
				"properties": {
					"to": {
						"type": "number",
						"format": "double"
					},
					"from": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object"
			},
			"CustomDimension": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"dimensionId": {
						"$ref": "#/components/schemas/FieldId"
					},
					"table": {
						"type": "string"
					},
					"binType": {
						"$ref": "#/components/schemas/BinType"
					},
					"binNumber": {
						"type": "number",
						"format": "double"
					},
					"binWidth": {
						"type": "number",
						"format": "double"
					},
					"customRange": {
						"items": {
							"$ref": "#/components/schemas/BinRange"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"dimensionId",
					"table",
					"binType"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Pick_CompiledDimension.label-or-name_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"label": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"label"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"MetricQuery": {
				"properties": {
					"metadata": {
						"properties": {
							"hasADateDimension": {
								"$ref": "#/components/schemas/Pick_CompiledDimension.label-or-name_"
							}
						},
						"required": [
							"hasADateDimension"
						],
						"type": "object"
					},
					"customDimensions": {
						"items": {
							"$ref": "#/components/schemas/CustomDimension"
						},
						"type": "array"
					},
					"additionalMetrics": {
						"items": {
							"$ref": "#/components/schemas/AdditionalMetric"
						},
						"type": "array"
					},
					"tableCalculations": {
						"items": {
							"$ref": "#/components/schemas/TableCalculation"
						},
						"type": "array"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"sorts": {
						"items": {
							"$ref": "#/components/schemas/SortField"
						},
						"type": "array"
					},
					"filters": {
						"$ref": "#/components/schemas/Filters"
					},
					"metrics": {
						"items": {
							"$ref": "#/components/schemas/FieldId"
						},
						"type": "array"
					},
					"dimensions": {
						"items": {
							"$ref": "#/components/schemas/FieldId"
						},
						"type": "array"
					},
					"exploreName": {
						"type": "string"
					}
				},
				"required": [
					"tableCalculations",
					"limit",
					"sorts",
					"filters",
					"metrics",
					"dimensions",
					"exploreName"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"session_cookie": {
				"type": "apiKey",
				"in": "cookie",
				"name": "connect.sid"
			},
			"api_key": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization",
				"description": "Value should be 'ApiKey <your key>'"
			}
		}
	},
	"info": {
		"title": "Lightdash API",
		"version": "0.1021.4",
		"description": "Open API documentation for all public Lightdash API endpoints.  # Authentication Before you get started, you might need to create a Personal Access Token to authenticate via the API. You can create a token by following this guide: https://docs.lightdash.com/references/personal_tokens\n",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Lightdash Support",
			"email": "support@lightdash.com",
			"url": "https://docs.lightdash.com/help-and-contact/contact/contact_info/"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/projects/{projectUuid}/explores": {
			"put": {
				"operationId": "SetExplores",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessEmpty"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorPayload"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "projectUuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {},
								"type": "array"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetExplores",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"results": {
											"$ref": "#/components/schemas/ApiExploresResults"
										},
										"status": {
											"type": "string",
											"enum": [
												"ok"
											],
											"nullable": false
										}
									},
									"required": [
										"results",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorPayload"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "projectUuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/projects/{projectUuid}/explores/{exploreId}": {
			"get": {
				"operationId": "GetExplore",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"results": {
											"$ref": "#/components/schemas/ApiExploreResults"
										},
										"status": {
											"type": "string",
											"enum": [
												"ok"
											],
											"nullable": false
										}
									},
									"required": [
										"results",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorPayload"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "exploreId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "projectUuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/projects/{projectUuid}/explores/{exploreId}/compileQuery": {
			"post": {
				"operationId": "CompileQuery",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"results": {
											"$ref": "#/components/schemas/ApiCompiledQueryResults"
										},
										"status": {
											"type": "string",
											"enum": [
												"ok"
											],
											"nullable": false
										}
									},
									"required": [
										"results",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorPayload"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "exploreId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "projectUuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MetricQuery"
							}
						}
					}
				}
			}
		},
		"/api/v1/projects/{projectUuid}/explores/compileCustomQuery": {
			"post": {
				"operationId": "CompileCustomQuery",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"results": {
											"$ref": "#/components/schemas/ApiCompiledQueryResults"
										},
										"status": {
											"type": "string",
											"enum": [
												"ok"
											],
											"nullable": false
										}
									},
									"required": [
										"results",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorPayload"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "projectUuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"explore": {
										"$ref": "#/components/schemas/Explore"
									},
									"metricQuery": {
										"$ref": "#/components/schemas/MetricQuery"
									}
								},
								"required": [
									"explore",
									"metricQuery"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/projects/{projectUuid}/explores/{exploreId}/downloadCsv": {
			"post": {
				"operationId": "DownloadCsvFromExplore",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"results": {
											"properties": {
												"jobId": {
													"type": "string"
												}
											},
											"required": [
												"jobId"
											],
											"type": "object"
										},
										"status": {
											"type": "string",
											"enum": [
												"ok"
											],
											"nullable": false
										}
									},
									"required": [
										"results",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorPayload"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "exploreId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "projectUuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/MetricQuery"
									},
									{
										"properties": {
											"hiddenFields": {
												"items": {
													"type": "string"
												},
												"type": "array"
											},
											"columnOrder": {
												"items": {
													"type": "string"
												},
												"type": "array"
											},
											"customLabels": {
												"properties": {},
												"additionalProperties": {
													"type": "string"
												},
												"type": "object"
											},
											"showTableNames": {
												"type": "boolean"
											},
											"csvLimit": {
												"type": "number",
												"format": "double",
												"nullable": true
											},
											"onlyRaw": {
												"type": "boolean"
											}
										},
										"required": [
											"columnOrder",
											"showTableNames",
											"onlyRaw"
										],
										"type": "object"
									}
								]
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}